# Vectorize c and c++ code