--- a/Earthfile
+++ b/Earthfile
@@ -59,7 +59,7 @@ c:
   RUN apk add --no-cache gcc build-base
 
   COPY ./src/leibniz.c ./
-  RUN --no-cache gcc leibniz.c -o leibniz -O3 -s -static -flto -ffast-math -march=native -mtune=native -fomit-frame-pointer
+  RUN --no-cache gcc leibniz.c -o leibniz -O3 -s -static -flto -march=native -mtune=native -fomit-frame-pointer -fno-signed-zeros -fno-trapping-math -fassociative-math
   RUN --no-cache ./scbench "./leibniz" -i $iterations -l "gcc --version" --export json --lang "C (gcc)"
   SAVE ARTIFACT ./scbench-summary.json AS LOCAL ./results/c.json
 
@@ -90,7 +90,7 @@ cpp:
   RUN apk add --no-cache gcc build-base
 
   COPY ./src/leibniz.cpp ./
-  RUN --no-cache g++ leibniz.cpp -o leibniz -O3 -s -static -flto -ffast-math -march=native -mtune=native -fomit-frame-pointer
+  RUN --no-cache g++ leibniz.cpp -o leibniz -O3 -s -static -flto -march=native -mtune=native -fomit-frame-pointer -fno-signed-zeros -fno-trapping-math -fassociative-math
   RUN --no-cache ./scbench "./leibniz" -i $iterations -l "g++ --version" --export json --lang "C++ (g++)"
   SAVE ARTIFACT ./scbench-summary.json AS LOCAL ./results/cpp.json