Adding `-ffast-math` compilation option and moving the calculation of variable `x` inside the scope of the for-loop allows gcc to vectorize the calculation for C and C++. This results in a significant speedup on my machine, going from around 100ms to 50ms. The change does not significantly affect the accuracy of the result. Analyses below

<details>
<summary>C</summary>

#### Original
```json
{
    "Started": "18-10-2022 19:33:20",
    "Ended": "18-10-2022 19:33:20",
    "Language": "C (gcc)",
    "Version": "11.2.1",
    "Command": "./leibniz",
    "Accuracy": 0.7222222222222222,
    "CalculatedPi": "3.1415926635893259",
    "Iterations": 3,
    "Average": "93.971662ms"
}
```
#### Proposed
```json
{
    "Started": "18-10-2022 19:56:48",
    "Ended": "18-10-2022 19:56:48",
    "Language": "C (gcc)",
    "Version": "11.2.1",
    "Command": "./leibniz",
    "Accuracy": 0.7222222222222222,
    "CalculatedPi": "3.1415926635857296",
    "Iterations": 3,
    "Average": "48.815929ms"
}
```
</details>

<details>
<summary>C++</summary>

#### Original
```json
{
    "Started": "18-10-2022 19:57:40",
    "Ended": "18-10-2022 19:57:41",
    "Language": "C++ (g++)",
    "Version": "11.2.1",
    "Command": "./leibniz",
    "Accuracy": 0.7222222222222222,
    "CalculatedPi": "3.1415926635893259",
    "Iterations": 3,
    "Average": "96.320308ms"
}
```
#### Proposed
```json
{
    "Started": "18-10-2022 19:59:43",
    "Ended": "18-10-2022 19:59:43",
    "Language": "C++ (g++)",
    "Version": "11.2.1",
    "Command": "./leibniz",
    "Accuracy": 0.7222222222222222,
    "CalculatedPi": "3.1415926635857296",
    "Iterations": 3,
    "Average": "49.448502ms"
}
```
</details>