@Moelf 2022-10-18
Fast math is waaay to dangerous probably not how usually people compile their application