@giordano 2022-10-18
Mandatory reminder that in general `-ffast-math` is bad:

* https://gcc.gnu.org/bugzilla/show_bug.cgi?id=55522
* https://github.com/llvm/llvm-project/issues/57589
* https://kristerw.github.io/2021/10/19/fast-math/
* https://simonbyrne.github.io/notes/fastmath/
* https://moyix.blogspot.com/2022/09/someones-been-messing-with-my-subnormals.html

You should be able to achieve the same effect with `-fassociative-math`, but GCC doesn't seem to like this flag for me:
```console
% gcc leibniz.c -o leibniz -O3 -s -static -flto -fassociative-math -march=native -mtune=native -fomit-frame-pointer
cc1: warning: ‘-fassociative-math’ disabled; other options take precedence
lto1: warning: ‘-fassociative-math’ disabled; other options take precedence
lto1: warning: ‘-fassociative-math’ disabled; other options take precedence
```
It works with Clang instead.
```c
#pragma clang fp reassociate(on)
```
would also work without any additional command line in Clang, I didn't find an equivalent pragma directive for GCC.