@LazyKernel 2022-10-19
Good point. I was not aware of the multitude of issues `-ffast-math` brings with it. Using just `-fassociative-math` seems like a much better option. For GCC though, we also need to enable `-fno-signed-zeros` and `-fno-trapping-math` (as per gcc manual). The combination `-fassociative-math -fno-signed-zeros -fno-trapping-math` produces practically the same result as above, with fewer downsides as far as I can see.